# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Isekai",
        "description": "Servicio REST que expone información de un mundo isekai ficticio.",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/v1/info/strata": {
            "get": {
                "tags": [
                    "Info"
                ],
                "summary": "Obtener listado de estratos",
                "operationId": "f7ff093a685070b9efe56634d733461a",
                "responses": {
                    "200": {
                        "description": "Listado de estratos obtenido exitosamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "code": {
                                                "type": "string",
                                                "example": "001"
                                            },
                                            "name": {
                                                "type": "string",
                                                "example": "Estrato Alto"
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "No hay información de estratos disponible",
                        "content": {
                            "application/problem+json": {
                                "schema": {
                                    "properties": {
                                        "type": {
                                            "type": "string",
                                            "example": "https://example.com/"
                                        },
                                        "title": {
                                            "type": "string",
                                            "example": "string"
                                        },
                                        "status": {
                                            "type": "integer",
                                            "example": 404
                                        },
                                        "detail": {
                                            "type": "string",
                                            "example": "string"
                                        },
                                        "instance": {
                                            "type": "string",
                                            "example": "https://example.com/"
                                        },
                                        "properties": {
                                            "type": "object",
                                            "example": {
                                                "additionalProp1": "string",
                                                "additionalProp2": "string",
                                                "additionalProp3": "string"
                                            },
                                            "additionalProperties": {
                                                "type": "string"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Error interno no manejado",
                        "content": {
                            "application/problem+json": {
                                "schema": {
                                    "properties": {
                                        "type": {
                                            "type": "string",
                                            "example": "https://example.com/"
                                        },
                                        "title": {
                                            "type": "string",
                                            "example": "string"
                                        },
                                        "status": {
                                            "type": "integer",
                                            "example": 500
                                        },
                                        "detail": {
                                            "type": "string",
                                            "example": "string"
                                        },
                                        "instance": {
                                            "type": "string",
                                            "example": "https://example.com/"
                                        },
                                        "properties": {
                                            "type": "object",
                                            "example": {
                                                "additionalProp1": "string",
                                                "additionalProp2": "string",
                                                "additionalProp3": "string"
                                            },
                                            "additionalProperties": {
                                                "type": "string"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        }
    },
...
}